UI Design
An interface for a fictional product development firm  "idea."  I designed. The function of the application is to showcase the firm's work in an organized and searchable manner. The firm's work is organized into three main libraries, Design, Marketing, and Engineering. The outcome is for prospective clients to easily search and view the firms' case studies, while also enabling members to openly critique and question the firms work. I hope to complete the rest of the interface views of this fictional product development firm  "idea.".